Combine the related codes in QueryBindeerTests
